fixed transition to register/loggedIn Pages
