+++ Frontend/Frontend/Pages/LoggedIn.cshtml
@@ -1,0 +1,27 @@
+﻿@page
+@model Frontend.Pages.LoggedInModel
+@{
+}
+<link rel="stylesheet" type="text/css" href="~/css/index.css" />
+
+
+<div id="mainDiv">
+        @foreach (var item in @Model.boardList)
+        {
+            <div>
+           
+                <h5 onclick="location.href='/Board?id=@item.Id'">@item.Title</h5>
+            
+            </div>
+            
+        }
+
+
+        <div id="newBoardForm">
+            <form onsubmit="CreateBoard()">
+                <input type="text" id="btitle" name="btitle" value="New board title">
+                <input type="submit" id="submit-btn" name="submit-btn" value="Create">
+            </form>
+    </div>
+
+</div>
